requirements.txt
fastapi
uvicorn
torch
torchvision
timm
transformers
neptune
numpy
pandas
flask
requests
pillow
opencv-python
python-multipart
